requirements.txt
python-decouple
langchain-core
langchain-google-genai
psycopg2-binary
numpy